% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uneFonction.R
\name{predict.CIps}
\alias{predict.CIps}
\title{A predict function for cips}
\usage{
\method{predict}{CIps}(object, newdata, X = NULL, y = NULL, ...)
}
\arguments{
\item{object}{An object of class 'cips'.}

\item{newdata}{A dataframe containing new data to make predictions.}

\item{X}{Explanatory variables, default is NULL.}

\item{y}{Target corresponding to the explanatory variables, default is NULL.}

\item{...}{Additional arguments for future use.}
}
\value{
A numeric vector of predicted values.
}
\description{
This function generates predictions based on a cips object.
}
\details{
When X and y are not NULL, they represent the data used to select the model.
These data are then reused by recalibrating on the selected subset to obtain the beta estimates (hybrid approach).
}
